ingestion/requirements.txt
psycopg2-binary
pandas
boto3
python-dotenv
requests